Move tests into suite subdirectories
