+++ Makefile
@@ -13,7 +13,33 @@ LIB_DIR := lib
 # Test runner
 TEST_DIR := tests
 TEST_BIN := $(TEST_DIR)/test-compiler
-TEST_SOURCES := $(TEST_DIR)/test_compiler.c $(TEST_DIR)/test_helpers.c $(TEST_DIR)/test_emitbc_header.c $(TEST_DIR)/test_emitbc_opcode_map.c $(TEST_DIR)/test_emitbc_jumps.c $(TEST_DIR)/test_emitbc_invariants.c $(TEST_DIR)/test_opcode_schema.c $(TEST_DIR)/test_pipeline_golden.c $(TEST_DIR)/test_pipeline_source_golden.c $(TEST_DIR)/test_ir_validate.c $(TEST_DIR)/test_pipeline_negative_matrix.c $(TEST_DIR)/test_absyn_lifecycle.c $(TEST_DIR)/test_semant.c $(TEST_DIR)/test_sdiss_fixtures.c $(TEST_DIR)/test_parser_examples_obj_golden.c $(TEST_DIR)/test_interpret_semantics_golden.c $(TEST_DIR)/test_fixture_policy.c $(TEST_DIR)/test_compiler_context_failures.c $(TEST_DIR)/test_compiler_diag_pipeline.c $(TEST_DIR)/test_parser_input_api.c $(TEST_DIR)/test_sys_compile_libcall.c $(TEST_DIR)/test_item_cache.c
+TEST_SHARED_SOURCES := \
+	$(TEST_DIR)/shared/test_compiler.c \
+	$(TEST_DIR)/shared/test_helpers.c \
+	$(TEST_DIR)/shared/test_fixture_policy.c
+TEST_CORE_SOURCES := \
+	$(TEST_DIR)/core/test_absyn_lifecycle.c \
+	$(TEST_DIR)/core/test_semant.c \
+	$(TEST_DIR)/core/test_ir_validate.c \
+	$(TEST_DIR)/core/test_opcode_schema.c \
+	$(TEST_DIR)/core/test_parser_input_api.c \
+	$(TEST_DIR)/core/test_item_cache.c
+TEST_COMPILER_SOURCES := \
+	$(TEST_DIR)/compiler/test_emitbc_header.c \
+	$(TEST_DIR)/compiler/test_emitbc_opcode_map.c \
+	$(TEST_DIR)/compiler/test_emitbc_jumps.c \
+	$(TEST_DIR)/compiler/test_emitbc_invariants.c \
+	$(TEST_DIR)/compiler/test_pipeline_golden.c \
+	$(TEST_DIR)/compiler/test_pipeline_source_golden.c \
+	$(TEST_DIR)/compiler/test_pipeline_negative_matrix.c \
+	$(TEST_DIR)/compiler/test_parser_examples_obj_golden.c \
+	$(TEST_DIR)/compiler/test_sdiss_fixtures.c \
+	$(TEST_DIR)/compiler/test_compiler_context_failures.c \
+	$(TEST_DIR)/compiler/test_compiler_diag_pipeline.c \
+	$(TEST_DIR)/compiler/test_sys_compile_libcall.c
+TEST_INTERPRETER_SOURCES := \
+	$(TEST_DIR)/interpreter/test_interpret_semantics_golden.c
+TEST_SOURCES := $(TEST_SHARED_SOURCES) $(TEST_CORE_SOURCES) $(TEST_COMPILER_SOURCES) $(TEST_INTERPRETER_SOURCES)
 
 
 # Library of shared functions
@@ -94,7 +120,7 @@ test: $(TEST_BIN)
 	./$(TEST_BIN)
 
 $(TEST_BIN): $(TEST_SOURCES) $(LIB) scomp sdiss sin
-	$(CC) $(CFLAGS) $(DEBUG) -Isrc -o $@ $(TEST_SOURCES) $(LIB) $(LDFLAGS) $(LIBS)
+	$(CC) $(CFLAGS) $(DEBUG) -Isrc -I$(TEST_DIR) -o $@ $(TEST_SOURCES) $(LIB) $(LDFLAGS) $(LIBS)
 
 clean:
 	rm -rf $(OBJ_DIR)/*.o $(OBJ_DIR)/*.d $(LIB) $(LIB_DIR) \
