+++ tests/compiler/test_compiler_diag_pipeline.c
