+++ tests/compiler/test_emitbc_header.c
