+++ tests/compiler/test_emitbc_invariants.c
