+++ tests/compiler/test_emitbc_opcode_map.c
