+++ tests/compiler/test_parser_examples_obj_golden.c
