+++ tests/compiler/test_pipeline_golden.c
