+++ tests/compiler/test_pipeline_source_golden.c
