+++ tests/core/test_absyn_lifecycle.c
