+++ tests/core/test_ir_validate.c
