+++ tests/core/test_item_cache.c
