+++ tests/core/test_opcode_schema.c
