+++ tests/core/test_semant.c
