+++ tests/interpreter/test_interpret_semantics_golden.c
